requirements.txt
langchain
python-dotenv
google-cloud-aiplatform